fixing rename package error
